+++ src/tensor_network_editor/__init__.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
     from .analysis import analyze_contraction, analyze_spec
     from .builder import IndexHandle, NetworkBuilder, TensorHandle
     from .canonicalization import canonicalize_spec
-    from .editor import EditorLaunchOptions, open_editor
+    from .editor import EditorLaunchOptions, EditorThemeName, open_editor
     from .internal.diffing._diffing import diff_specs, semantic_diff_specs
     from .io import PythonLoadOptions, load_python_spec, load_spec, save_spec
     from .linting import lint_spec
@@ -73,6 +73,7 @@ __all__ = [
     "EdgeEndpointRef",
     "EdgeSpec",
     "EditorLaunchOptions",
+    "EditorThemeName",
     "EditorResult",
     "EngineName",
     "DotRenderOptions",
@@ -124,6 +125,7 @@ _LAZY_EXPORTS: dict[str, str] = {
     "EdgeEndpointRef": ".models",
     "EdgeSpec": ".models",
     "EditorLaunchOptions": ".editor",
+    "EditorThemeName": ".editor",
     "EditorResult": ".models",
     "EngineName": ".models",
     "DotRenderOptions": ".rendering",